feat: 이슈 전체 목록에 카운터 값 추가

+++ BE/src/main/java/team01/issuetracker/repository/IssueRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 
 @Repository
 public interface IssueRepository extends CrudRepository<Issue, Long> {
-    long countByIsOpen(boolean state);
+    int countByIsOpen(boolean state);
 
     int countByIsOpenAndMilestoneId(boolean state, Long id);
